refactor(ffi): split lib.rs (4054 lines) into topic modules
Break up the 145-function FFI surface into types / auth / historical /
streaming / utility / error / panic sub-modules. Pure move — all extern
"C" symbols and #[repr(C)] layouts preserved byte-for-byte. ABI
unchanged; downstream C/C++/Go/Node consumers see zero difference.
